launch-agent/pyproject.toml
[project]
name = "robarkham-launch-agent"
version = "0.1.0"
description = "Launch agent for robcotech.pro deployments"
authors = [
    { name = "Codex" },
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "asyncpg>=0.30.0,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "google-adk>=1.15.0,<2.0.0",
    "google-auth>=2.38.0,<3.0.0",
    "google-cloud-aiplatform[evaluation]>=1.130.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "uvicorn>=0.34.0,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.agents-cli]
name = "robarkham-launch-agent"
description = "Launch coordinator for robcotech.pro"
base_template = "adk"
agent_directory = "app"

[tool.agents-cli.create_params]
deployment_target = "none"
session_type = "in_memory"
cicd_runner = "skip"
include_data_ingestion = false
datastore = "none"
